requirements.txt
web3==6.11.3
pandas==2.1.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
rich==13.7.0
requests==2.31.0
scikit-learn==1.3.2
